Refactor commandtexthistory and update DG
