+++ src/main/java/seedu/address/ui/MainWindow.java
@@ -2,9 +2,12 @@ package seedu.address.ui;
 
 import static seedu.address.logic.commands.HistoryCommand.MESSAGE_SHOW_HISTORY_SUCCESS;
 
+import java.util.function.Consumer;
+import java.util.function.Function;
 import java.util.logging.Logger;
 
 import javafx.event.ActionEvent;
+import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.scene.control.MenuItem;
 import javafx.scene.control.TextInputControl;
@@ -125,7 +128,8 @@ public class MainWindow extends UiPart<Stage> {
         StatusBarFooter statusBarFooter = new StatusBarFooter(logic.getAddressBookFilePath());
         statusbarPlaceholder.getChildren().add(statusBarFooter.getRoot());
 
-        CommandBox commandBox = new CommandBox(this::executeCommand);
+        CommandBox commandBox = new CommandBox(this::executeCommand, logic::getPreviousCommandTextFromHistory,
+                logic::getNextCommandTextFromHistory);
         commandBoxPlaceholder.getChildren().add(commandBox.getRoot());
 
         callHistoryPanel = new CallHistoryPanel();
