+++ src/test/java/seedu/address/logic/commands/CommandTextHistoryTest.java
@@ -3,6 +3,7 @@ package seedu.address.logic.commands;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
+import seedu.address.model.CommandTextHistory;
 
 /**
  * Contains unit tests for CommandTextHistory.
